Fix new transaction_manager tests
